+++ src/test/java/domain/ladder/FloorTest.java
@@ -1,9 +1,7 @@
-package domain;
+package domain.ladder;
 
 import common.exception.message.ExceptionMessage;
-import domain.bridge.strategy.BridgeGeneratorStub;
-import domain.ladder.Floor;
-import domain.ladder.LadderBridge;
+import domain.ladder.strategy.BridgeGeneratorStub;
 import java.util.List;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
