+++ src/test/java/domain/ladder/LadderTest.java
@@ -1,12 +1,8 @@
-package domain;
+package domain.ladder;
 
 import static org.junit.jupiter.api.Assertions.assertAll;
 
-import domain.bridge.strategy.BridgeGeneratorStub;
-import domain.ladder.Floor;
-import domain.ladder.Ladder;
-import domain.ladder.LadderBridge;
-import domain.ladder.LadderHeight;
+import domain.ladder.strategy.BridgeGeneratorStub;
 import domain.player.PlayerName;
 import domain.player.PlayerNames;
 import java.util.List;
