+++ src/test/java/domain/ladder/strategy/BridgeGeneratorStub.java
@@ -1,4 +1,4 @@
-package domain.bridge.strategy;
+package domain.ladder.strategy;
 
 import domain.ladder.LadderBridge;
 import domain.ladder.strategy.BridgeGenerator;
