+++ src/test/java/domain/player/PlayerNameTest.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.player;
 
 import common.exception.message.ExceptionMessage;
 import domain.player.PlayerName;
